Geoip: Remove previous C99 compatibility changes with v100
Since we're compiling against v120_xp now, C99 is supported, which makes the previous compatibility changes pointless.
This reverts to the state of the original content.
For maintainability reason, it's welcomed.

For reference, the only change now is the moving of PACKAGE_VERSION define to be available for linux too (since we don't use configure.ac)