Improve set_task tasks timing and performance
Half-rewrite of task system. The highlights are improved performance due to use of std::vector instead of a linked list, which is faster due to locality, and ordering tasks by execution time so it's not necessary to iterate over the entire task list every frame.

In addition, code is now more modern (C++11 stuff) and should be safer.

Tasks before would be executed in batches every 0.1 seconds, I guess so as to avoid checking every frame if a certain task was due or not.
This however leads to bad performance and many issues.

Bad performance because
- If there are many tasks, instead of them executing in a spread fashion, they will all or most execute together in a single frame. This might lead to the frame not meeting it's allocated time (if sys_ticrate 1000, if a single frame takes more than 1ms this will result in server fps drop of course)
- Because of the limitation and poor timing of tasks (can't have a 0.05 task, or a 0.15 task will be really imprecise), it's a widespread recommendation/workaround to use entities and hook think. This method is obviously slower and has more overhead than a plain task, but was the only valid alternative often.

Because of this, although it might seem counterintuitive, checking for due tasks every frame actually improves performance, since it's all about not having any slow frames and not average CPU usage (besides the fact that it takes negligible CPU time)

Besides, it fixes the following issues
*Need to use entities because of the known limitations
*Huge timing drifts (imagine how much will a repetitive 0.15s task drift over time).

I've tested it with a monster size plugin where the slightest compatibility issue would arise, and seems to be fine.