Workaround for incorrect heap deallocation on conditional operator
Fixes #621.

Using conditional operator can produce different heap allocation size in different branches (one of them can also produce no allocation). The problem is that heap deallocation is executed unconditionally for both branches. It is even incorrect if both branches use the same allocation size. This workaround fixes this problem by adding an additional heap allocation for a branch with a minimal allocation size (or without an allocation). It also fixes a problem when allocation sizes are the same.